created app payment and payment model moved inside it
